refactor(app): isolate delivery wiring tests
